feat: add #[derive(KeyCodec)] for kvapi key structs
Manual KeyCodec impls are mechanical: encode each named field via
KeyCodec::encode_key in declaration order, decode in the same order.
The new derive in databend-meta-kvapi-derive generates exactly that,
preserving the existing wire format byte-for-byte.

Macro path resolution goes through proc-macro-crate so a renamed dep
(`mykvapi = { package = "databend-meta-kvapi" }`) works transparently;
reaching the macro through a transitive re-exporter is not supported
on purpose, to keep the resolution rule simple.

Two ways to opt a field out of percent-escaping:

- `Raw` — a `repr(transparent)` newtype around `String` whose
  KeyCodec impl uses push_raw / next_raw. Usable directly as a field
  type.
- `#[key_codec(raw)]` on a `String` field — the derive routes that
  field through `Raw`'s impl via the zero-copy `Raw::from_ref`.

`#[key_codec(raw)]` on a non-String field fails to compile (the
generated `Raw::from_ref(&self.field)` yields a `&String`/`&u64`
mismatch). Unknown sub-options like `#[key_codec(rwa)]` produce an
explicit error rather than silently behaving as "not raw". Enums,
tuple structs, unit structs, and unions are rejected up front.
trybuild snapshots in tests/ui lock in each error.

Also impl KeyCodec for u32: wire format matches u64 (decimal segment),
but decode rejects values that overflow u32 instead of silently
truncating — needed for keys mixing u64 table ids with u32 column ids.
