+++ Cargo.toml
@@ -49,7 +49,7 @@ databend-meta-version = { path = "crates/common/version" }
 base2histogram = "0.2.3"
 databend-base = "0.4.0"
 map-api = "0.4.2"
-openraft = { version = "=0.10.0-alpha.18", features = ["serde", "tracing-log", "runtime-stats"] }
+openraft = { version = "=0.10.0-alpha.19", features = ["serde", "tracing-log", "runtime-stats"] }
 sled = { version = "0.34", default-features = false }
 state-machine-api = "0.3.4"
 
@@ -101,6 +101,7 @@ seq-marked = { version = "0.3.5", features = ["seq-marked-serde", "seq-marked-bi
 serde = { version = "1.0.164", features = ["derive", "rc"] }
 serde_json = { version = "1.0.85", default-features = false, features = ["preserve_order", "unbounded_depth"] }
 stream-more = "0.1.3"
+structkey = "0.1.0"
 strum = "0.24.1"
 strum_macros = "0.24"
 tempfile = "3.4.0"
